Created WPShapes class(consistent naming), removed WPRelationships class, comments cleanup
