+++ modules/org.pathvisio.gui/src/org/pathvisio/gui/CommonActions.java
@@ -57,7 +57,8 @@ import org.pathvisio.core.view.Template;
 import org.pathvisio.core.view.VPathway;
 import org.pathvisio.core.view.VPathwayElement;
 import org.pathvisio.core.view.ViewActions;
-import org.pathvisio.core.view.WPRelationships;
+import org.pathvisio.core.view.WPShapes;
+import org.pathvisio.core.view.WPShapes;
 import org.pathvisio.gui.dialogs.AboutDlg;
 import org.pathvisio.gui.dialogs.PathwayElementDialog;
 import org.pathvisio.gui.dialogs.PublicationXRefDialog;
@@ -297,31 +298,31 @@ public class CommonActions implements ApplicationEventListener {
 	// actions for "WPs Relationships (aka interactions)" section
 		 newWPInteractionActions = new Action[] {
 				 new NewElementAction(e, new DefaultTemplates.LineTemplate(
-						 "Conversion", LineStyle.SOLID, LineType.LINE, WPRelationships.WP_CONVERSION, ConnectorType.STRAIGHT) //Metabolic Reaction Relationship
+						 "Conversion", LineStyle.SOLID, LineType.LINE, WPShapes.WP_CONVERSION, ConnectorType.STRAIGHT) //Metabolic Reaction Relationship
 				 ),
 				 new NewElementAction(e, new DefaultTemplates.LineTemplate(
-						 "Catalysis", LineStyle.SOLID, LineType.LINE, WPRelationships.WP_CATALYSIS, ConnectorType.STRAIGHT) //Catalysis Relationship
+						 "Catalysis", LineStyle.SOLID, LineType.LINE, WPShapes.WP_CATALYSIS, ConnectorType.STRAIGHT) //Catalysis Relationship
 				 ),
 				 new NewElementAction(e, new DefaultTemplates.LineTemplate(
-						 "Inhibition", LineStyle.SOLID, LineType.LINE, WPRelationships.WP_INHIBITION, ConnectorType.STRAIGHT) //Inhibiting Relationship
+						 "Inhibition", LineStyle.SOLID, LineType.LINE, WPShapes.WP_INHIBITION, ConnectorType.STRAIGHT) //Inhibiting Relationship
 				 ),
 				 new NewElementAction(e, new DefaultTemplates.LineTemplate(
-						 "Post Translational Modification", LineStyle.SOLID, LineType.LINE, WPRelationships.WP_MODIFICATION, ConnectorType.STRAIGHT) //PMT Relationship
+						 "Post Translational Modification", LineStyle.SOLID, LineType.LINE, WPShapes.WP_MODIFICATION, ConnectorType.STRAIGHT) //PMT Relationship
 				 ),
 				 new NewElementAction(e, new DefaultTemplates.LineTemplate(
 						 "Basic Directed", LineStyle.DASHED, LineType.LINE, LineType.ARROW, ConnectorType.STRAIGHT) // Basic Directed Relationship (reusing Basic interaction panel)
 				 ),
 				 new NewElementAction(e, new DefaultTemplates.LineTemplate(
-						 "Stimulation", LineStyle.SOLID, LineType.LINE, WPRelationships.WP_STIMULATION, ConnectorType.STRAIGHT) // Stimulation Relationship
+						 "Stimulation", LineStyle.SOLID, LineType.LINE, WPShapes.WP_STIMULATION, ConnectorType.STRAIGHT) // Stimulation Relationship
 				 ),
 				 new NewElementAction(e, new DefaultTemplates.LineTemplate(
-						 "Binding", LineStyle.SOLID, LineType.LINE, WPRelationships.WP_BINDING, ConnectorType.STRAIGHT // Binding Relationship
+						 "Binding", LineStyle.SOLID, LineType.LINE, WPShapes.WP_BINDING, ConnectorType.STRAIGHT // Binding Relationship
 				 )),
 				 new NewElementAction(e, new DefaultTemplates.LineTemplate(
-						 "Transcription-Translation", LineStyle.SOLID, LineType.LINE, WPRelationships.WP_TRANSLATION, ConnectorType.STRAIGHT // Transcription/Translation Relationship
+						 "Transcription-Translation", LineStyle.SOLID, LineType.LINE, WPShapes.WP_TRANSLATION, ConnectorType.STRAIGHT // Transcription/Translation Relationship
 				 )),
 				 new NewElementAction(e, new DefaultTemplates.LineTemplate(
-						 "Translocation", LineStyle.SOLID, LineType.LINE, WPRelationships.WP_TRANSLOCATION, ConnectorType.STRAIGHT // Translocation (cellular movement) Relationship
+						 "Translocation", LineStyle.SOLID, LineType.LINE, WPShapes.WP_TRANSLOCATION, ConnectorType.STRAIGHT // Translocation (cellular movement) Relationship
 				 )),
 				 new NewElementAction(e, new DefaultTemplates.LineTemplate(
 						 "Basic Undirected", LineStyle.SOLID, LineType.LINE, LineType.LINE, ConnectorType.STRAIGHT // Basic UNDirected Relationship (reusing Basic interaction panel)
